requirements_blood_network.txt
Flask==2.3.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3